Fix an issue where compilation may fail on some non-x64 platforms

+++ src/util/frame/audio.rs
@@ -69,7 +69,7 @@ impl Audio {
     #[inline]
     pub fn set_channel_layout(&mut self, value: ChannelLayout) {
         unsafe {
-            (*self.as_mut_ptr()).channel_layout = value.bits();
+            (*self.as_mut_ptr()).channel_layout = value.bits() as u64;
         }
     }
 
@@ -140,7 +140,7 @@ impl Audio {
             panic!("out of bounds");
         }
 
-        if !<T as Sample>::is_valid(self.format(), self.channels()) {
+        if !<T as Sample>::is_valid(self.format(), self.channels() as u16) {
             panic!("unsupported type");
         }
 
@@ -153,7 +153,7 @@ impl Audio {
             panic!("out of bounds");
         }
 
-        if !<T as Sample>::is_valid(self.format(), self.channels()) {
+        if !<T as Sample>::is_valid(self.format(), self.channels() as u16) {
             panic!("unsupported type");
         }
